advanced-agent/pyproject.toml
[project]
name = "advanced-agent"
version = "0.1.0"
description = "AI Research Agent for Developer Tools"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.1.1",
    "aiohttp>=3.9.1",
    "beautifulsoup4>=4.12.2",
    "lxml>=4.9.3",
    "langchain>=0.3.26",
    "langchain-core>=0.3.26",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.2",
    "brotli>=1.0.9",
    "pydantic>=2.11.7",
    # New dependencies for enhanced functionality
    "redis>=5.0.0",              # For caching
    "structlog>=23.0.0",         # For structured logging
    "tenacity>=8.2.0",           # For retry logic
    "httpx>=0.25.0",             # Alternative HTTP client
    "asyncio-throttle>=1.0.0",   # For rate limiting
    "python-multipart>=0.0.6",   # For file uploads
    "websockets>=12.0",          # For real-time updates
]